feat: start redoing the find script in python
The idea is pretty good, but I don't know if I'll do it. The bash
script is slow but it works pretty well. This is frustrating to
work with but for a real-world use its not a problem as it will
only be run like once or twice a month. Maybe when the site is up
and running. Right now I should focus back on strace... This is the
project I was supposed to do if I remember correctly^^ I guess I
got a bit carried away on this one (it's not the first time lol).