docs: move check tasks to MAYBE
They are actually hard to do. For the life of me, I cant seem to find
a unique and coherent naming convention for the diverse array of
processor architectures that exist... As for the checking of the list
themselves... Well this is kinda the same thing. I should ask someone
who is actually knowledgeable on this subject.