fix: do not use last_run in Makefile
Rename it to last_update since that's what it really represents. Only
update it in the update.bash script (makes sense doesnt it ?), when
the last version of the linux kernel sources have been fetched.

This makes more sense and makes and not having the date changes
polluting the diffs makes it easier to change and test the build
scripts. The Makefile is also clearer like that.