feat: handle the sigreturn as a special case
The find_syscalls script is going to look for the syscall in the
assembly files. If it finds it a hard-coded prototype is provided
for the csv files. I dont like to hard-code stuff but it seems to
make sense in this case since the prototype is basically empty (no
parameters and standard return) and that it is strictly the same
across all the architectures it was missing on.

There's a lot less missing syscalls now, so thats pretty cool.
However sigreturn is still missing on i386. This is because the
syscall is not actually implemented in assembly. It is found in a
C file but the definition is not taken into account by the script
(investigate this). Maybe this can be fixed generally, if it does
not change anything else, or it will have to be extra-hard-coded.