feat: add syscalls to find by asm
The sparc 32's execv should not really be in this list. The problem
is that it's one of these syscalls that can only be found in assembly
files but that actually has parameters (actually "syscall" is kind of
the same since it at least needs the NR...). So here we have syscalls
that are implemented and found but that mistakingly do not list any
parameters. That probably should be a new status in the CSV files.

The good news is that now there is only one syscall missing.