feat: parse 32bit specific syscall defines
This successfully gets rid of the last missing syscall!

However it also reveals a previously unknown error in the i386 syscall
table. The fanotify_mark syscall has two definitions, including a 32bit
specific one. It was previously ignored by the script so it wrongfully
used the default implementation. Now it gets the specific one but it
cannot parse it. This is because of marcro in the parameters of the
define. It splits a 64bit value in two for the 32bit version.

This is a bit tricky because this should be done at the preprocessing
step of the script. It requires the Kconfig file and the endianess of
the current architecture. So now the script has been changed so that
previously fatal prototype parsing errors are now admitted. It sets
the 'noparam' status, the same as for the asm syscalls. It makes sense
because this is really the same error. The script is not complicated
enough to handle theses edge cases.