docs: remove last mandatory thing TODO
This is beautiful! The project can now pass into the 'finished' category.
For me it means that it essentially works (unti proven otherwise of course
hahaha). From now on the work on this will mainly be maintenance and bug
fixing. When I have time I can do the 'maybe' tasks in the TODO file, but
that's all bonus stuff :D