refactor: remove linux submodule
Now the linux kernel is cloned as a dependency to the Makefile's
last_update rule. If it already exists it simply pulls the last version
of the master branch and creates the last_update file. Basically this
simplifies the process by quite a lot. All that's required to build and
update the app is a 'make re' command. Then the running docker instance
simply needs to be rebuilt with docker compose.

Now the linux/ kernel repository is what it should be, data. This is less
clear at the repository level that this project depends on it but this is
simpler to execute and maintain. Now the updates do not require a new
commit and the repository only consists of files actually doing the build.